+++ IntegrationTests/BraintreeUnionPay_IntegrationTests.m
@@ -24,7 +24,10 @@
 
 - (void)pendFetchCapabilities_returnsCardCapabilities {
     XCTestExpectation *expectation = [self expectationWithDescription:@"Callback invoked"];
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
     [self.cardClient fetchCapabilities:@"6212345678901232" completion:^(BTCardCapabilities * _Nullable cardCapabilities, NSError * _Nullable error) {
+#pragma clang diagnostic pop
         XCTAssertNil(error);
         XCTAssertFalse(cardCapabilities.isDebit);
         XCTAssertTrue(cardCapabilities.isUnionPay);
@@ -104,7 +107,10 @@
     [self waitForExpectationsWithTimeout:5 handler:nil];
 
     expectation = [self expectationWithDescription:@"Callback invoked"];
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
     [self.cardClient tokenizeCard:request options:nil completion:^(BTCardNonce * _Nullable tokenizedCard, NSError * _Nullable error) {
+#pragma clang diagnostic pop
         XCTAssertNil(error);
         XCTAssertTrue([tokenizedCard.nonce isANonce]);
         XCTAssertEqual(tokenizedCard.cardNetwork, BTCardNetworkUnionPay);