+++ Sources/BraintreeCard/Public/BraintreeCard/BTCardClient.h
@@ -68,7 +68,8 @@ typedef NS_ENUM(NSInteger, BTCardClientErrorType) {
 */
 - (void)tokenizeCard:(BTCardRequest *)request
              options:(nullable NSDictionary *)options
-          completion:(void (^)(BTCardNonce * _Nullable tokenizedCard, NSError * _Nullable error))completion;
+          completion:(void (^)(BTCardNonce * _Nullable tokenizedCard, NSError * _Nullable error))completion
+DEPRECATED_MSG_ATTRIBUTE("The UnionPay SMS integration is deprecated, as UnionPay can now be processed as a credit card through their partnership with Discover. Use `BTCardClient.tokenizeCard(card: completion:)`.");
 
 @end