+++ Sources/BraintreeDataCollector/Public/BraintreeDataCollector/BTConfiguration+DataCollector.h
@@ -12,11 +12,13 @@
 /**
  Indicates whether Kount is enabled for the merchant account.
 */
-@property (nonatomic, readonly, assign) BOOL isKountEnabled;
+@property (nonatomic, readonly, assign) BOOL isKountEnabled
+DEPRECATED_MSG_ATTRIBUTE("Kount Custom support will be removed in the next major version. Use `PPDataCollector.collectPayPalDeviceData()`.");
 
 /**
  Returns the Kount merchant id set in the Gateway
 */
-@property (nonatomic, readonly, assign) NSString *kountMerchantID;
+@property (nonatomic, readonly, assign) NSString *kountMerchantID
+DEPRECATED_MSG_ATTRIBUTE("Kount Custom support will be removed in the next major version. Use `PPDataCollector.collectPayPalDeviceData()`.");
 
 @end