+++ Sources/BraintreeUnionPay/Public/BraintreeUnionPay/BTConfiguration+UnionPay.h
@@ -7,6 +7,7 @@
 /**
  BTConfiguration category for UnionPay
  */
+DEPRECATED_MSG_ATTRIBUTE("The UnionPay SMS integration is deprecated, as UnionPay can now be processed as a credit card through their partnership with Discover. Use `BTCardClient.tokenizeCard(card: completion:)`.")
 @interface BTConfiguration (UnionPay)
 
 /**